some changes

+++ b_lib/DATABASE.py
@@ -2,12 +2,14 @@ import logging
 import time
 import json
 import base64
+import pandas as pd
 import psycopg2
 from CONFIG import Config as cfg
 import logging
 from datetime import datetime
 from psycopg2.extras import Json, RealDictCursor
 import os
+
 
 class DateBase:
     def __init__(self):
@@ -46,6 +48,13 @@ class DateBase:
         :return:
         """
         query = f"SELECT НОМЕР_ЗАПРОСА FROM {cfg.table_tasks} WHERE {column} = '{value}'"
+        with self.connect():
+            self.cursor.execute(query)
+            task = self.cursor.fetchone()[0]
+        return task
+
+    def get_status_task(self, theme):
+        query = f"SELECT СТАТУС FROM {cfg.table_tasks} WHERE ТЕМА_ПИСЬМА = '{theme}'"
         with self.connect():
             self.cursor.execute(query)
             task = self.cursor.fetchone()[0]
@@ -91,7 +100,7 @@ class DateBase:
                 params = f"{params} {column} = '{value}',"
         query = f"UPDATE {table} SET{params} ДАТА_ОБРАБОТКИ = '{time_processing}' WHERE НОМЕР_ЗАПРОСА = '{tasks}'"
         with self.connect():
-            #только для случаев обновления ДАТЫ _ОБРАБОТКИ
+            # только для случаев обновления ДАТЫ _ОБРАБОТКИ
             if isinstance(tasks, list):
                 for task in tasks:
                     query = f"UPDATE {table} SET{params} " \
@@ -115,11 +124,13 @@ class DateBase:
         }
         :return:
         """
+        time_processing = datetime.now()
         with self.connect():
             self.cursor.execute(
-                f"UPDATE {cfg.table_tasks} SET СПИСОК_РАССЫЛКИ = %s "
+                f"UPDATE {cfg.table_tasks} SET СПИСОК_РАССЫЛКИ = %s, "
+                f"ОТПРАВЛЕНО = %s "
                 "WHERE НОМЕР_ЗАПРОСА = %s",
-                (Json(data), task_id,)
+                (Json(data), time_processing, task_id,)
             )
 
     def add_project_db(self, task, project, date_project, list_sogl, link, status):
@@ -141,7 +152,6 @@ class DateBase:
             )
             logging.info('Запись в таблицу EOSDO_427 внесена')
 
-
     def add_task_db(self, tasks):
         """
         Добавить новую запись в БД
@@ -156,6 +166,7 @@ class DateBase:
                 organization = task['body']['organization']
                 initiator = task['body']['initiator']
                 request = task
+                task_life = task['body']['task_life']
                 time_processing = datetime.now()
                 self.cursor.execute(
                     f"INSERT INTO {cfg.table_tasks} "
@@ -164,15 +175,16 @@ class DateBase:
                     "ОРГАНИЗАЦИЯ, "
                     "ИНИЦИАТОР, "
                     "ЗАПРОС, "
+                    "СРОК_ДЕЙСТВИЯ, "
                     "ДАТА_ОБРАБОТКИ, "
                     "СТАТУС "
                     ") "
-                    "VALUES (%s, %s, %s, %s, %s, %s)",
-                    (task_id, organization, initiator, Json(request), time_processing, status,)
+                    "VALUES (%s, %s, %s, %s, %s, %s, %s)",
+                    (task_id, organization, initiator, Json(request), task_life, time_processing, status,)
                 )
         logging.info('Новые задания добавлены в БД')
 
-    def get_new_tasks(self):
+    def get_new_tasks(self) -> list:
         """
         Получение записей со статусом Новое
         :return: Возвращает НОМЕР_ЗАПРОСА в формате [(1,), (2,)]
@@ -187,25 +199,45 @@ class DateBase:
 
     def clean_db(self):
         """
-        Удаление из тиблицы tasks_427 записей со статусом Закрыт и датой обработки более cfg.last_day дней
-        Удаление из тиблицы responses_427 записей c датой обработки более cfg.last_day дней
+        Удаление из таблицы tasks_427 записей со истекшим сроком действия ТКП
+        Удаление из таблицы responses_427 записей c датой обработки более cfg.last_day дней
         :return:
         """
+        from b_lib.b_excel import Excel
         date_now = datetime.now()
         with self.connect():
-            query = f"DELETE FROM {cfg.table_tasks} WHERE " \
-                    f"EXTRACT(day FROM '{date_now}' - ДАТА_ОБРАБОТКИ) >= {cfg.last_day} AND СТАТУС = 'Закрыт'"
+            # если истек срок действия ТКП, закрываем задание
+            query = "UPDATE TASKS_427 SET СТАТУС = 'Закрыт' WHERE " \
+                    f"СРОК_ДЕЙСТВИЯ < DATE_PART('day', '{date_now}' - ОТПРАВЛЕНО) AND СТАТУС = 'Отправлено'"
             self.cursor.execute(query)
+            # Задания для удаления из таблицы TASKS
+            query = f"SELECT * FROM {cfg.table_tasks} WHERE " \
+                    f"DATE_PART('day', '{date_now}' - ДАТА_ОБРАБОТКИ) >= {cfg.last_day} AND " \
+                    f"СТАТУС = 'Закрыт'"
+            df = pd.read_sql(query, self.connection)
+            if df.empty is False:
+                # Записываем в файл
+                Excel().write_to_excel(cfg.removed_tasks, df)
+                tasks = df['НОМЕР_ЗАПРОСА'].tolist()
+                # Задания для удаления из таблицы eosdo [task1, task2] превращаем в (task1, task2)
+                tasks_set = str(tasks).replace('[', '(').replace(']', ')')
+                query = f"SELECT * FROM {cfg.table_eosdo} WHERE НОМЕР_ЗАПРОСА IN {tasks_set}"
+                df = pd.read_sql(query, self.connection)
+                # Записываем их в файл
+                Excel().write_to_excel(cfg.removed_eosdo, df)
+                # Удаляем данные из БД
+                query = f"DELETE FROM {cfg.table_tasks} WHERE НОМЕР_ЗАПРОСА IN {tasks_set}"
+                self.cursor.execute(query)
 
-            query = f"DELETE FROM {cfg.table_responses} WHERE " \
-                    f"EXTRACT(day FROM '{date_now}' - ОТПРАВЛЕНО) >= {cfg.last_day}"
-            self.cursor.execute(query)
-
-            query = f"DELETE FROM {cfg.table_eosdo} WHERE " \
-                    f"EXTRACT(day FROM '{date_now}' - ДАТА_ОБРАБОТКИ) >= {cfg.last_day}" \
-                    f"AND СТАТУС_ЕОСДО = 'Закрыт' OR СТАТУС_ЕОСДО = 'Доработка'"
-
-            self.cursor.execute(query)
+            # Задания для удаления из таблицы RESPONSES
+            query = f"SELECT * FROM RESPONSES_427 WHERE " \
+                    f"DATE_PART('day', '{date_now}' - ОТПРАВЛЕНО) >= {cfg.last_day_response}"
+            df = pd.read_sql(query, self.connection)
+            if df.empty is False:
+                Excel().write_to_excel(cfg.removed_responses, df)
+                query = f"DELETE FROM {cfg.table_responses} WHERE " \
+                        f"EXTRACT(day FROM '{date_now}' - ОТПРАВЛЕНО) >= {cfg.last_day_response}"
+                self.cursor.execute(query)
 
     def remove_task(self, task):
         with self.connect():
@@ -225,21 +257,15 @@ class DateBase:
         """
 
         date_now = datetime.now()
+        root_query = f"SELECT НОМЕР_ЗАПРОСА, ОРГАНИЗАЦИЯ FROM {cfg.table_tasks} WHERE " \
+                     f"EXTRACT(epoch FROM age('{date_now}', ДАТА_ОБРАБОТКИ))/60 >= {cfg.check_min} " \
+                     f"AND СТАТУС = 'Мониторинг' " \
+                     "AND ОШИБКИ IS NULL "
         with self.connect():
-            if organization is not None:
-                query = f"SELECT НОМЕР_ЗАПРОСА FROM {cfg.table_tasks} WHERE " \
-                        f"EXTRACT(epoch FROM age('{date_now}', ДАТА_ОБРАБОТКИ))/60 >= {cfg.check_min} " \
-                        f"AND СТАТУС = 'Мониторинг' " \
-                        f"AND ОРГАНИЗАЦИЯ = '{organization}'" \
-                        f"AND ОШИБКИ IS NULL"
-
-            elif organization is None:
-                query = f"SELECT НОМЕР_ЗАПРОСА, ОРГАНИЗАЦИЯ FROM {cfg.table_tasks} WHERE " \
-                        f"EXTRACT(epoch FROM age('{date_now}', ДАТА_ОБРАБОТКИ))/60 >= {cfg.check_min}" \
-                        f"AND СТАТУС = 'Мониторинг' " \
-                        f"AND ОШИБКИ IS NULL " \
-                        f"ORDER BY ОРГАНИЗАЦИЯ" \
-
+            if organization:
+                query = root_query + f"AND ОРГАНИЗАЦИЯ = '{organization}'"
+            else:
+                query = root_query + f"ORDER BY ОРГАНИЗАЦИЯ"
             self.cursor.execute(query)
             tasks = self.cursor.fetchall()
         return tasks
@@ -285,7 +311,9 @@ class DateBase:
                                  СПОСОБ_ОТПРАВКИ TEXT,
                                  ТЕМА_ПИСЬМА TEXT,
                                  СПИСОК_РАССЫЛКИ JSONB,
+                                 ОТПРАВЛЕНО TIMESTAMP,
                                  СТАТУС TEXT NOT NULL,
+                                 СРОК_ДЕЙСТВИЯ INTEGER NOT NULL,
                                  ДАТА_ОБРАБОТКИ TIMESTAMP NOT NULL,
                                  ОШИБКИ TEXT
                                  ); '''
@@ -294,7 +322,7 @@ class DateBase:
             self.cursor.execute(f'''CREATE INDEX TASKS_TABLE ON {cfg.table_tasks} 
                                     (ОРГАНИЗАЦИЯ, ИНИЦИАТОР, ДАТА_ОБРАБОТКИ, СТАТУС, 
                                     ОШИБКИ, СПОСОБ_ОТПРАВКИ,
-                                    СПИСОК_РАССЫЛКИ
+                                    СПИСОК_РАССЫЛКИ, СРОК_ДЕЙСТВИЯ
                                  ); '''
                                 )
 
@@ -304,7 +332,9 @@ class DateBase:
         """
         with self.connect():
             self.cursor.execute(f'''CREATE TABLE IF NOT EXISTS {cfg.table_eosdo}   
-                                 (НОМЕР_ЗАПРОСА VARCHAR PRIMARY KEY,
+                                 (
+                                 ID SERIAL PRIMARY KEY,
+                                 НОМЕР_ЗАПРОСА VARCHAR,
                                  НОМЕР_ПРОЕКТА TEXT,
                                  ДАТА_ПРОЕКТА DATE,
                                  ЛИСТ_СОГЛАСОВАНИЯ JSONB,
@@ -312,7 +342,8 @@ class DateBase:
                                  ДАТА_РЕГ DATE,
                                  СТАТУС_ЕОСДО TEXT,
                                  ССЫЛКА TEXT,
-                                 ДАТА_ОБРАБОТКИ TIMESTAMP NOT NULL
+                                 ДАТА_ОБРАБОТКИ TIMESTAMP NOT NULL,
+                                 FOREIGN KEY (НОМЕР_ЗАПРОСА) REFERENCES TASKS_427 (НОМЕР_ЗАПРОСА) ON DELETE CASCADE
                                  ); '''
                                 )
 
@@ -361,46 +392,28 @@ class DateBase:
 
     def get_sending_tasks(self, organization=None):
         """
-        Получение записей из БД со статусом Отправлено, способами отправки 'сешанный' или 'еосдо' и датой обработки
-        более 120 мин
-        :param organization: имя организации
-        :return:
-        """
+            Получение записей из БД со статусом Отправлено, способами отправки 'сешанный' или 'еосдо' и датой обработки больше 60 мин
+           :param organization: имя организации
+           :return:
+            """
         date_now = datetime.now()
+        root_query = "SELECT НОМЕР_ЗАПРОСА, ОРГАНИЗАЦИЯ " \
+                     f"FROM {cfg.table_tasks} " \
+                     f"WHERE EXTRACT(epoch FROM age('{date_now}', ДАТА_ОБРАБОТКИ))/60 >= {cfg.check_min} " \
+                     "AND СТАТУС = 'Отправлено' " \
+                     "AND (СПОСОБ_ОТПРАВКИ = 'смешанный' OR СПОСОБ_ОТПРАВКИ = 'еосдо') " \
+                     "AND ОШИБКИ IS NULL "
+
         with self.connect() as conn:
             with conn.cursor(cursor_factory=psycopg2.extras.RealDictCursor):
                 if organization:
-                    self.cursor.execute(
-                        "SELECT НОМЕР_ЗАПРОСА "
-                        f"FROM {cfg.table_tasks} "
-                        "WHERE "
-                        "EXTRACT(epoch FROM age(%s, ДАТА_ОБРАБОТКИ))/120 >= %s"
-                        "AND "
-                        "СТАТУС = 'Отправлено'"
-                        "AND "
-                        "ОРГАНИЗАЦИЯ = %s "
-                        "AND "
-                        "(СПОСОБ_ОТПРАВКИ = 'смешанный' OR СПОСОБ_ОТПРАВКИ = 'еосдо') "
-                        "AND ОШИБКИ IS NULL",
-                        (date_now, cfg.check_min, organization,)
-                    )
+                    query = root_query + f"AND ОРГАНИЗАЦИЯ = '{organization}'"
+                    self.cursor.execute(query)
                     tasks = [task[0] for task in self.cursor.fetchall()]
 
-                elif organization is None:
-                    self.cursor.execute(
-                        "SELECT НОМЕР_ЗАПРОСА, ОРГАНИЗАЦИЯ "
-                        f"FROM {cfg.table_tasks} "
-                        "WHERE "
-                        "EXTRACT(epoch FROM age(%s, ДАТА_ОБРАБОТКИ))/60 >= %s"
-                        "AND "
-                        "СТАТУС = 'Отправлено'"
-                        "AND "
-                        "(СПОСОБ_ОТПРАВКИ = 'смешанный' OR СПОСОБ_ОТПРАВКИ = 'еосдо') "
-                        "AND ОШИБКИ IS NULL "
-                        "ORDER BY ОРГАНИЗАЦИЯ",
-
-                        (date_now, cfg.check_min,)
-                    )
+                else:
+                    query = root_query + 'ORDER BY ОРГАНИЗАЦИЯ'
+                    self.cursor.execute(query)
                     tasks = self.cursor.fetchall()
         return tasks
 
@@ -417,3 +430,43 @@ class DateBase:
             self.cursor.execute(f"DROP TABLE {cfg.table_eosdo};")
 
 
+if __name__ == '__main__':
+    dict = {
+        'a': 'b'
+    }
+
+    db = DateBase()
+    db.clean_db()
+    # db.drop_table_eosdo()
+    # db.create_table_tasks()
+    # db.create_table_eosdo()
+    # # db.drop_table_tasks()
+    #
+    # # db.drop_table_respond()
+    # db.create_table_respond()
+
+    # db.create_table_tasks()
+
+    # db.response_db('93GnYRx', {'data':'data'}, 'ПРОЕКТ_СОЗДАН',  1000333)
+    # db.add_reg_project('199-80', '22-9.2/18', '24.11.2022')
+    # db.get_new_tasks()
+    # db.drop_table()
+    # db.create_table()
+
+    # db.get_one('1000333', 'ЗАПРОС', 'dict')
+    # db.get_task('РЕГ_НОМЕР', '22-9.2/25')
+
+    # db.get_sogl_db('199000')
+    # db.add_sogl_db('199000', data)
+    # for key in [{'organization': 'АО"Гринатом"', 'task_id': '434f-567'},
+    #             {'organization': 'АО"Атомэнергопроект"', 'task_id': '434f-411'},
+    #             {'organization': 'АО "АСЭ"', 'task_id': '434f-4980000'},
+    #             {'organization': 'АО "ТВЭЛ"', 'task_id': '434f-4777'}
+    #             ]:
+    #     db.add_task(key['task_id'], key['organization'])
+    # db.get_tasks_monitoring("АО \"Гринатом\"")
+    # body_task = db.get_one('1000333', 'ЗАПРОС', 'dict')
+    # files_task = body_task["files"]
+    # for file in files_task:
+    #     with open(os.path.join(cfg.saved_files, file['file_name']), "wb") as newfile:
+    #         newfile.write(base64.b64decode(file["file"]))
